tidy: move test_pytest to _ast
